refactor: extract constants, signals, create components for profile.tsx
