+++ src/modules/theme/constants.ts
@@ -1,0 +1,2 @@
+export const CARD_BACKGROUND_COLOR = 'bg-slate-800'
+export const CARD_STYLE = 'mt-5 pt-5 rounded-lg'
